Add annotation and path validation skip to attach
