+++ cmd/oras/file.go
@@ -17,7 +17,9 @@ package main
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"os"
 	"path/filepath"
 
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
@@ -55,4 +57,14 @@ func loadFiles(ctx context.Context, store *file.Store, annotations map[string]ma
 		fmt.Println("Uploading empty artifact")
 	}
 	return files, nil
+}
+
+// decodeJSON decodes a json file v to filename.
+func decodeJSON(filename string, v interface{}) error {
+	file, err := os.Open(filename)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	return json.NewDecoder(file).Decode(v)
 }
